Add "map restore" command to reload the last-saved map.
Sometimes the automapper screws up, inserts a room where no room should be, and shifts half the Red Desert to kingdom come. This is a useful way of dealing with the problem.

When mapping, you should "map save" frequently to save the map to $mudlethome/map.dat. If the map becomes corrupted, you can use the "map restore" command to read back map.dat, overwritting the broken map in memory.

Added an echo to the "map save" command, paralleling that in "map restore". IMO it's useful to see where the map is being saved in case you want to make additional backups.